Move generic FindFeatureFlag recipe to search package too
