+++ src/test/java/org/openrewrite/featureflags/search/FindFeatureFlagTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.openrewrite.featureflags;
+package org.openrewrite.featureflags.search;
 
 import org.junit.jupiter.api.Test;
 import org.openrewrite.DocumentExample;
